weight/hx711py/requirement.txt
RPi.GPIO
hx711py
numpy
requests
matplotlib
pillow
statistics
counter
picamera2
tk
boto3
openai


